+++ DICOM/DicomDatasetExtensions.cs
@@ -1,5 +1,8 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Linq;
+
+using Dicom.IO.Writer;
 
 namespace Dicom {
 	public static class DicomDatasetExtensions {