+++ DICOM/DicomElement.cs
@@ -142,7 +142,7 @@ namespace Dicom {
 				return (T)DicomParseable.Parse<T>(_values[item]);
 
 			if (typeof(T).IsEnum)
-				return (T)Enum.Parse(typeof(T), _values[item]);
+				return (T)Enum.Parse(typeof(T), _values[item], true);
 
 			throw new InvalidCastException("Unable to convert DICOM " + ValueRepresentation.Code + " value to '" + typeof(T).Name + "'");
 		}
@@ -261,7 +261,7 @@ namespace Dicom {
 					throw new ArgumentOutOfRangeException("item", "Index is outside the range of available value items");
 
 				var s = ByteConverter.Get<Tv>(Buffer, item).ToString();
-				return (T)Enum.Parse(typeof(T), s);
+				return (T)Enum.Parse(typeof(T), s, true);
 			}
 
 			if (typeof(T).IsValueType) {
@@ -850,7 +850,7 @@ namespace Dicom {
 		#endregion
 
 		#region Public Members
-		public override T Get<T>(int item = -1) {
+		public override T Get<T>(int item = 0) {
 			if (typeof(T) == typeof(int) || typeof(T) == typeof(int[]))
 				return (T)(object)base.Get<T>(item);
 
@@ -949,6 +949,9 @@ namespace Dicom {
 	/// <summary>Unique Identifier (UI)</summary>
 	public class DicomUniqueIdentifier : DicomMultiStringElement {
 		#region Public Constructors
+		public DicomUniqueIdentifier(DicomTag tag, params string[] values) : base(tag, values) {
+		}
+
 		public DicomUniqueIdentifier(DicomTag tag, params DicomUID[] values) : base(tag, values.Select(x => x.UID).ToArray()) {
 		}
 
@@ -1044,7 +1047,7 @@ namespace Dicom {
 		#endregion
 
 		#region Public Members
-		public override T Get<T>(int item = -1) {
+		public override T Get<T>(int item = 0) {
 			if (typeof(T) == typeof(int) || typeof(T) == typeof(int[]))
 				return (T)(object)base.Get<T>(item);