+++ DICOM/IO/FileReference.cs
@@ -2,10 +2,19 @@
 using System.IO;
 
 namespace Dicom.IO {
-	public sealed class FileReference : IDisposable {
+	public sealed class FileReference {
 		public FileReference(string fileName, bool isTempFile = false) {
 			Name = fileName;
 			IsTempFile = isTempFile;
+		}
+
+		~FileReference() {
+			if (IsTempFile) {
+				try {
+					Delete();
+				} catch {
+				}
+			}
 		}
 
 		public string Name {
@@ -13,7 +22,7 @@ namespace Dicom.IO {
 			private set;
 		}
 
-		/// <summary>File will be deleted when object is Disposed.</summary>
+		/// <summary>File will be deleted when object is <c>Disposed</c>.</summary>
 		public bool IsTempFile {
 			get;
 			internal set;
@@ -34,6 +43,22 @@ namespace Dicom.IO {
 				File.Delete(Name);
 		}
 
+		/// <summary>
+		/// Moves file and updates internal reference.
+		/// 
+		/// Calling this method will also remove set the <see cref="IsTempFile"/> property to <c>False</c>.
+		/// </summary>
+		/// <param name="dstFileName"></param>
+		public void Move(string dstFileName, bool overwrite = false) {
+			// delete if overwriting; let File.Move thow IOException if not
+			if (File.Exists(dstFileName) && overwrite)
+				File.Delete(dstFileName);
+
+			File.Move(Name, dstFileName);
+			Name = Path.GetFullPath(dstFileName);
+			IsTempFile = false;
+		}
+
 		public byte[] GetByteRange(int offset, int count) {
 			byte[] buffer = new byte[count];
 
@@ -43,15 +68,6 @@ namespace Dicom.IO {
 			}
 
 			return buffer;
-		}
-
-		public void Dispose() {
-			if (IsTempFile) {
-				try {
-					Delete();
-				} catch {
-				}
-			}
 		}
 	}
 }